+++ checkov/common/checks_infra/solvers/complex_solvers/base_complex_solver.py
@@ -3,10 +3,8 @@ from __future__ import annotations
 from abc import abstractmethod
 from typing import List, Any, Tuple, Dict, TYPE_CHECKING, Optional
 
-from igraph import Graph
 from networkx import DiGraph
 
-from checkov.common.graph.checks_infra import debug
 from checkov.common.graph.checks_infra.enums import SolverType
 from checkov.common.graph.checks_infra.solvers.base_solver import BaseSolver
 
@@ -38,29 +36,8 @@ class BaseComplexSolver(BaseSolver):
         passed_vertices = []
         failed_vertices = []
         unknown_vertices = []
-        if isinstance(graph_connector, Graph):
-            select_kwargs = {}
-            if self.resource_types:
-                select_kwargs = {"resource_type_in": self.resource_types}
 
-            for data in graph_connector.vs.select(**select_kwargs)["attr"]:
-                result = self.get_operation(data)
-                if result is None:
-                    unknown_vertices.append(data)
-                elif result:
-                    passed_vertices.append(data)
-                else:
-                    failed_vertices.append(data)
-
-            debug.complex_connection_block(
-                solvers=self.solvers,
-                operator=self.operator,
-                passed_resources=passed_vertices,
-                failed_resources=failed_vertices,
-            )
-
-            return passed_vertices, failed_vertices, unknown_vertices
-        elif isinstance(graph_connector, DiGraph):
+        if isinstance(graph_connector, DiGraph):
             for _, data in graph_connector.nodes(data=True):
                 if self.resource_type_pred(data, self.resource_types):
                     result = self.get_operation(data)
@@ -72,6 +49,7 @@ class BaseComplexSolver(BaseSolver):
                         failed_vertices.append(data)
             return passed_vertices, failed_vertices, unknown_vertices
 
+        # isinstance(graph_connector, PyDiGraph):
         for _, data in graph_connector.nodes():
             if self.resource_type_pred(data, self.resource_types):
                 result = self.get_operation(data)