+++ checkov/terraform/checks/resource/aws/AppsyncAPICacheEncryptionInTransit.py
@@ -4,7 +4,7 @@ from checkov.common.models.enums import CheckCategories, CheckResult
 
 class AppsyncCacheEncryptionInTransit(BaseResourceValueCheck):
     def __init__(self) -> None:
-        name = "Ensure Appsync API Cache is encrypted in transit"
+        name = "Ensure AppSync API Cache is encrypted in transit"
         id = "CKV_AWS_215"
         supported_resources = ["aws_appsync_api_cache"]
         categories = [CheckCategories.ENCRYPTION]